Conflicts:
	CMakeLists.txt

+++ src/jdb.cpp
@@ -18,13 +18,12 @@
  ***/
 
 #include "jdb.hpp"
-#include <msl/shader_loader.hpp>
 #include "game_states/main_menu.hpp"
 
 namespace jdb {
 
-const GLfloat WINDOW_WIDTH = 1024;
-const GLfloat WINDOW_HEIGHT = 768;
+const int WINDOW_WIDTH = 1024;
+const int WINDOW_HEIGHT = 768;
 
 
 /***
